Move files into slap subdirectory
