+++ src/slap/errors.c
