+++ src/slap/errors.h
