+++ src/slap/linalg.c
