+++ src/slap/linalg.h
