+++ src/slap/macros.h
