+++ src/slap/matrix.c
