+++ src/slap/matrix_checks.c
