+++ src/slap/matrix_checks.h
