+++ src/slap/printing.c
