+++ src/slap/submatrix.c
